+++ Sources/PointFree/Emails/FreeEpisodeEmail.swift
@@ -8,6 +8,7 @@ import HttpPipeline
 import HttpPipelineHtmlSupport
 import Models
 import Optics
+import PointFreeRouter
 import Prelude
 import Styleguide
 import View
