+++ Sources/PointFree/Emails/InviteEmail.swift
@@ -3,6 +3,7 @@ import FunctionalCss
 import Html
 import HtmlCssSupport
 import Models
+import PointFreeRouter
 import Prelude
 import Styleguide
 import View
