+++ Sources/PointFree/Episode/Episode.swift
@@ -1,4 +1,5 @@
 import Models
+import PointFreeRouter
 
 extension Episode {
   public var previousEpisodes: [Episode] {
