+++ Sources/PointFree/RouteHelpers.swift
@@ -1,9 +1,9 @@
-import PointFreeRouter
-
-public func path(to route: Route) -> String {
-  return pointFreeRouter.path(to: route)
-}
-
-public func url(to route: Route) -> String {
-  return pointFreeRouter.url(to: route)
-}
+//import PointFreeRouter
+//
+//public func path(to route: Route) -> String {
+//  return pointFreeRouter.path(to: route)
+//}
+//
+//public func url(to route: Route) -> String {
+//  return pointFreeRouter.url(to: route)
+//}
