+++ Sources/PointFree/StripeWebhooks.swift
@@ -9,6 +9,7 @@ import HttpPipelineHtmlSupport
 import Mailgun
 import Models
 import PointFreePrelude
+import PointFreeRouter
 import Prelude
 import Styleguide
 import Stripe
