+++ Sources/PointFreeRouter/PointFreeRouter.swift
@@ -33,3 +33,11 @@ public struct PointFreeRouter {
 }
 
 public var pointFreeRouter = PointFreeRouter()
+
+public func path(to route: Route) -> String {
+  return pointFreeRouter.path(to: route)
+}
+
+public func url(to route: Route) -> String {
+  return pointFreeRouter.url(to: route)
+}
