+++ Tests/PointFreeTests/AuthTests.swift
@@ -4,6 +4,7 @@ import Html
 import HttpPipeline
 @testable import PointFree
 import PointFreePrelude
+import PointFreeRouter
 import PointFreeTestSupport
 import Prelude
 import Optics
